fix:修复puppeteer渲染图片问题，优化项目结构
