+++ package.json
@@ -4,7 +4,8 @@
   "version": "1.1.8",
   "contributors": [
     "N. Sun <nsun200512@outlook.com>",
-    "ZoruaFox <zorua@vip.qq.com>"
+    "ZoruaFox <zorua@vip.qq.com>",
+    "AwAjie <ajie@awajie.com>"
   ],
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
@@ -28,14 +29,8 @@
     "koishi-plugin-adapter-onebot": "^6.8.0"
   },
   "peerDependenciesMeta": {
-    "@koishijs/plugin-canvas": {
-      "optional": true
-    },
-    "@koishijs/plugin-puppeteer": {
-      "optional": true
-    }
   },
   "dependencies": {
-    "@satorijs/element": "^3.1.8"
+    "puppeteer-core": "24.16.2"
   }
 }
