+++ src/utils/puppeteer/template.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>ysyunhei</title>
+  <style>
+    #app {
+      width: fit-content;
+    }
+  </style>
+</head>
+<body>
+  <div id="app">${content}</div>
+</body>
+</html>
