+++ tsconfig.json
@@ -12,7 +12,6 @@
     "esModuleInterop": true,
     "moduleResolution": "bundler",
     "jsx": "react-jsx",
-    "jsxImportSource": "@satorijs/element",
     "types": [
       "node",
       "yml-register/types"
